requirements.txt
pandas>=1.5.0,<2.1.0
numpy>=1.21.0,<1.25.0
pymysql>=1.0.0
sqlalchemy>=1.4.0,<2.1.0
akshare>=1.16.0
schedule>=1.2.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.28.0
tushare>=1.2.0
